requirements.txt
click==6.6
ecdsa==0.13.3
Flask==1.0
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Naked==0.1.31
nexmo==2.0.0
ordereddict==1.1
pycrypto==2.6.1
python-jose==1.2.0
PyYAML==5.1
requests==2.20.0
shellescape==3.4.1
six==1.10.0
Werkzeug==0.15.3
python-dotenv==0.7.1
